# 🚀 CI/CD Pipeline Performance Optimizations: 40-60% Faster Builds