--- a/.claude-flow/metrics/performance.json
+++ b/.claude-flow/metrics/performance.json
@@ -1,5 +1,5 @@
 {
-  "startTime": 1754498407222,
+  "startTime": 1754499885044,
   "totalTasks": 1,
   "successfulTasks": 1,
   "failedTasks": 0,