--- a/.claude-flow/metrics/task-metrics.json
+++ b/.claude-flow/metrics/task-metrics.json
@@ -1,10 +1,10 @@
 [
   {
-    "id": "cmd-hooks-1754498407282",
+    "id": "cmd-hooks-1754499740208",
     "type": "hooks",
     "success": true,
-    "duration": 61.46834899999999,
-    "timestamp": 1754498407344,
+    "duration": 71.112224,
+    "timestamp": 1754499740279,
     "metadata": {}
   }
 ]