… .claude-flow/metrics/system-metrics.json
… .github/pipeline-config.yml
… .github/pipeline-optimization.md
… .github/workflows/cache-fallback-warming.yml
… .github/workflows/cache-optimization.yml
… .github/workflows/cache-performance-monitor.yml
… .github/workflows/cache-strategies.yml
… .github/workflows/intelligent-load-balancer.yml
… .github/workflows/optimized-ci-cache.yml
… .github/workflows/optimized-ci.yml
… .github/workflows/optimized-pipeline-architecture.yml
… .github/workflows/parallel-optimization.yml
… .github/workflows/parallel-test-execution.yml
… CACHE_OPTIMIZATION_SUMMARY.md
… CI_CD_OPTIMIZATION_GUIDE.md
… PARALLEL_EXECUTION_OPTIMIZATION.md
… PIPELINE_ARCHITECTURE_SUMMARY.md
… src/plugins/traits.rs